pyproject.toml
# [build-system]
requires = ["setuptools", "wheel"]
build-backend = "flit_core.buildapi"

[project]
name = "omnisealbench"
description = "Watermarking benchmark evaluation tool"
readme = {file = "README.md", content-type = "text/markdown"}
authors = [{name = "Facebook AI Research"}]
requires-python = ">=3.10"
dynamic = ["version"]

classifiers=[
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering",
    "Topic :: Multimedia :: Image",
    "Topic :: Multimedia :: Video",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Multimedia :: Watermarking",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Development Status :: 4 - Beta",
]

dependencies = [
    "torch",
    "torchmetrics",
    "omegaconf",
    "pandas",
    "tqdm",
    "requests",
]

[project.urls]
  Source = "https://github.com/facebookresearch/omnisealbench"
  Tracker = "https://github.com/facebookresearch/omnisealbench/issues"

[project.scripts]
omniseal = "omnisealbench.cli:main"

[project.optional-dependencies]
dev = [
    "flake8",
    "mypy",
    "pytest",
    "black",
    "isort",
    "flake8",
    "pre-commit",
    "rich",
]
audio = [
    "soundfile",
    "datasets<4.0",
    "numpy<2",
    "pesq",
    # torch 2.9 and newer includes torchaudio 2.9 won't work yet for some audio attacks
    "torch<2.9",
    "torchaudio<2.9",
    # for torchaudio 2.9 and newer, we need torchcodec
    "torchcodec",
    "julius",
    "pystoi",
    "librosa<0.10",
    "matplotlib",
]
image = [
    "numpy<2",
    "torchvision",
    "scikit-learn",
    "scikit-image",
    "lpips",
    "kornia",
    "augly",
]
video = [
    "numpy<2",
    "torchvision",
    "scikit-learn",
    "scikit-image",
    "lpips",
    "kornia",
    "av",
    "ffmpeg-python",
    "pytorch-msssim",
]
leaderboard = [
    "gunicorn",
    "flask~=3.0.3",
    "werkzeug~=3.0.3",
    "watchdog",
    "flask-cors",
]

all = [
    "torch<2.9",
    "torchmetrics",
    "omegaconf",
    "pandas",
    "tqdm",
    "requests",

    "rich",
    "fire",

    # Audio profile
    "soundfile",
    "datasets<4.0",
    "numpy<2",
    "pesq",
    "torchaudio<2.9",
    "torchcodec",
    "julius",
    "pystoi",
    "librosa<0.10",
    "matplotlib",

    # Image profile
    "torchvision<0.24",
    "scikit-learn",
    "scikit-image",
    "lpips",
    "kornia",
    "augly",

    # Video profile
    "av",
    "ffmpeg-python",
    "pytorch-msssim",

    # Leaderboard (kept from existing)
    "gunicorn",
    "flask~=3.0.3",
    "werkzeug~=3.0.3",
    "watchdog",
    "flask-cors",
]

[tool.setuptools.dynamic]
version = {attr = "omnisealbench.__version__"}

[tool.setuptools.package-data]
"omnisealbench" = [
    "py.typed",
    "attacks/*.yaml",
    "cards/**/*.yaml",
    "configs/*.yaml",
    "datasets/**/*.yaml",
    "models/**/*.yaml",
    "attacks/*.yaml",
]

#####################################################################
#
# Linter format
# 
#####################################################################

[tool.mypy]
python_version = "3.11"
disable_error_code = "type-abstract,typeddict-unknown-key"
show_error_codes = true
check_untyped_defs = true
ignore_missing_imports = true
implicit_optional = true
implicit_reexport = true


[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "wam_src",
    "notebooks",
]

# Same as Black.
line-length = 100 # 88 (black)
indent-width = 4

# Assume Python 3.9
target-version = "py39"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F", "I"]
ignore = ["F403", "E731", "F401"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"